Re #896 Seems Working mex binning code without MPI
